Refactor: idea 관련 dto 패키지 수정
